requirements.txt
# Semantic Kernel for Python
semantic-kernel>=1.37.0

# Web framework
flask>=3.0.0

# Additional utilities
python-dotenv>=1.0.0
